+++ tests/test_initialization.py
@@ -9,7 +9,8 @@ from .test_classes import (MockPrognostic,
                            MockPrognosticWithExtraQuantities,
                            MockPrognosticWithMalformedExtraQuantities,
                            MockPrognosticWithExtraDimensionsAndSigmaLevels,
-                           MockPrognosticWithExtraDimensionsIn2d)
+                           MockPrognosticWithExtraDimensionsIn2d,
+                           MockImplicitWithAllAttributes)
 
 
 def test_no_components():
@@ -288,7 +289,7 @@ def test_basic_2d_coordinates():
     assert np.all(state['x'].values == random_x_values)
     assert np.all(state['y'].values == random_y_values)
 
-    for quantity in dummy.inputs:
+    for quantity in dummy._climt_inputs.keys():
         assert 'shore' in state[quantity].coords
         assert 'latitude' in state[quantity].coords
         assert state[quantity].coords['shore'].units is 'km'
@@ -313,6 +314,13 @@ def test_2d_coordinates_in_extra_dimensions():
     with pytest.raises(NotImplementedError) as excinfo:
         get_default_state([dummy])
     assert 'not yet supported' in str(excinfo.value)
+
+
+def test_create_implicit_diagnostics():
+
+    dummy = MockImplicitWithAllAttributes()
+
+    state = get_default_state([dummy])
 
 
 if __name__ == '__main__':